+++ react/src/app/society_bureau/router.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Switch, Route} from 'react-router-dom';
 import Dashboard from "./container/Dashboard";
-import SocietyListContainer from "./container/SocietyListContainer";
+import SocietyListContainer from "./society/containers/SocietyListContainer";
 import CreditContainer from "./credit/containers/CreditContainer";
 
 export default function SBRouter({ match }) {
