+++ react/src/app/society_bureau/society/containers/SocietyListContainer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import AdminContentWrapper from '../../../shared/admin_content_wrapper/AdminContentWrapper';
+import AdminContentWrapper from '../../../../shared/admin_content_wrapper/AdminContentWrapper';
 import SocietyTable from '../components/SocietyTable'
 
 class SocietyListContainer extends React.Component {
